Refactored folder names to match JS conventions | Fixes failing tests on Linux (letter case mismatch)
