+++ parsers/keyword-nodes/kwnodefun.js
@@ -1,7 +1,7 @@
 const constants = require("../../constants.js");
 const BaseNode = require("../basenode.js");
 const kwNodeJeki =  require("./kwnodejeki.js");
-const bracketExpressionNl = require("../nodeLiterals/bracketexpressionnl.js");
+const bracketExpressionNl = require("../node-literals/bracketexpressionnl.js");
 
 class KwNodeFun extends BaseNode {
