+++ parsers/node-literals/notoperatornl.js
