+++ tests/interpreters/inodegetti.test.js
@@ -6,7 +6,7 @@ const MainInterpreter = require("../../interpreters/maininterpreter.js");
 const iNodeGetJeki = require("../../interpreters/inodegetjeki.js");
 const Environment = require("../../environment.js");
 const iNodeTi = require("../../interpreters/inodejeki.js");
-const kwNodeTi = require("../../parsers/keywordnodes/kwnodejeki.js");
+const kwNodeTi = require("../../parsers/keyword-nodes/kwnodejeki.js");
 const Parser = require("../../parsers/parser.js");
 const Lexer = require("../../lexer.js");
 const InputStream = require("../../inputstream.js");
