+++ tests/interpreters/inodelthanoreq.test.js
@@ -5,7 +5,7 @@ jest.mock('fs', () => ({
 const MainInterpreter = require("../../interpreters/maininterpreter.js");
 const Environment = require("../../environment.js");
 const iNodeLthanOrEq = require("../../interpreters/inodelthanoreq.js");
-const kwNodeTi = require("../../parsers/keywordnodes/kwnodejeki.js");
+const kwNodeTi = require("../../parsers/keyword-nodes/kwnodejeki.js");
 const Parser = require("../../parsers/parser.js");
 const Lexer = require("../../lexer.js");
 const InputStream = require("../../inputstream.js");
