+++ tests/interpreters/inodepada.test.js
@@ -4,7 +4,7 @@ jest.mock('fs', () => ({
 
 const MainInterpreter = require("../../interpreters/maininterpreter.js");
 const iNodePada = require("../../interpreters/inodepada.js");
-const kwNodePada = require("../../parsers/keywordnodes/kwnodepada.js");
+const kwNodePada = require("../../parsers/keyword-nodes/kwnodepada.js");
 const Parser = require("../../parsers/parser.js");
 const lexer = require("../../lexer.js");
 const InputStream = require("../../inputstream.js");
