+++ tests/interpreters/inodesope.test.js
@@ -5,7 +5,7 @@ jest.mock('fs', () => ({
 const MainInterpreter = require("../../interpreters/maininterpreter.js");
 const Environment = require("../../environment.js");
 const iNodeSope = require("../../interpreters/inodesope.js");
-const kwNodeSope = require("../../parsers/keywordnodes/kwnodesope.js");
+const kwNodeSope = require("../../parsers/keyword-nodes/kwnodesope.js");
 const Parser = require("../../parsers/parser.js");
 const lexer = require("../../lexer.js");
 const InputStream = require("../../inputstream.js");
