+++ tests/parsers/keyword-nodes/kwnodegbewole.test.js
@@ -2,7 +2,7 @@ jest.mock('fs', () => ({
     readFileSync: jest.fn()
 }));
 
-const KwNodeGbeWole = require("../../../parsers/keywordnodes/kwnodegbewole.js");
+const KwNodeGbeWole = require("../../../parsers/keyword-nodes/kwnodegbewole.js");
 const Parser = require("../../../parsers/parser.js");
 const lexer = require("../../../lexer.js");
 const InputStream = require("../../../inputstream.js");
