+++ tests/parsers/keyword-nodes/kwnodeise.test.js
@@ -2,7 +2,7 @@ jest.mock('fs', () => ({
     readFileSync: jest.fn()
 }));
 
-const kwNodeIse = require("../../../parsers/keywordnodes/kwnodeise.js");
+const kwNodeIse = require("../../../parsers/keyword-nodes/kwnodeise.js");
 const Parser = require("../../../parsers/parser.js");
 const lexer = require("../../../lexer.js");
 const InputStream = require("../../../inputstream.js");
