+++ tests/parsers/keyword-nodes/kwnodesope.test.js
@@ -2,7 +2,7 @@ jest.mock('fs', () => ({
     readFileSync: jest.fn()
 }));
 
-const kwNodeSope = require("../../../parsers/keywordnodes/kwnodesope.js");
+const kwNodeSope = require("../../../parsers/keyword-nodes/kwnodesope.js");
 const Parser = require("../../../parsers/parser.js");
 const lexer = require("../../../lexer.js");
 const InputStream = require("../../../inputstream.js");
