+++ tests/parsers/keyword-nodes/kwnodeyi.test.js
@@ -2,7 +2,7 @@ jest.mock('fs', () => ({
     readFileSync: jest.fn()
 }));
 
-const kwNodeYi = require("../../../parsers/keywordnodes/kwnodeyi.js");
+const kwNodeYi = require("../../../parsers/keyword-nodes/kwnodeyi.js");
 const Parser = require("../../../parsers/parser.js");
 const Lexer = require("../../../lexer.js");
 const InputStream = require("../../../inputstream.js");
