+++ tests/parsers/node-literals/bracketexpressionnl.test.js
@@ -2,7 +2,7 @@ jest.mock('fs', () => ({
     readFileSync: jest.fn()
 }));
 
-const bracketExpNl = require("../../../parsers/nodeliterals/bracketexpressionnl.js");
+const bracketExpNl = require("../../../parsers/node-literals/bracketexpressionnl.js");
 const Parser = require("../../../parsers/parser.js");
 const Lexer = require("../../../lexer.js");
 const InputStream = require("../../../inputstream.js");
